make local single tts and stt runnables
