+++ .github/workflows/jt_tv_audio_processing.yml
@@ -1,13 +1,9 @@
 name: Process ORTM News to audio summary
 
 on:
-  push:
-    branches:
-      - '**'
-#on:
-#  schedule:
-#    # Set execution hour and minutes in UTC (a.k.a GMT)
-#    - cron: '30 23 * * *'
+  schedule:
+    # Set execution hour and minutes in UTC (a.k.a GMT)
+    - cron: '30 05 * * *'
 
 jobs:
   process_audio:
@@ -49,7 +45,7 @@ jobs:
        PLAYLIST_URL="https://youtube.com/playlist?list=PLDBQmURq6pOfBKc6WU0wXTg2vxAjxjQel&si=n9iNX7AUi-SpNN_N"
        
        # Search for the latest video inside the playlist
-       VIDEO_ID=$(yt-dlp --downloader aria2c --username oauth2 --password '' --flat-playlist -J "$PLAYLIST_URL" | jq -r '.entries[-2].id')
+       VIDEO_ID=$(yt-dlp --downloader aria2c --username oauth2 --password '' --flat-playlist -J "$PLAYLIST_URL" | jq -r '.entries[-1].id')
        if [ -z "$VIDEO_ID" ]; then
          echo "No video found in the playlist."
          exit 1
