--- service.yaml
@@ -1,74 +1,0 @@
-# for more info, see : https://github.com/GoogleCloudPlatform/cloud-run-samples/tree/main/multi-container/hello-nginx-sample
-apiVersion: serving.knative.dev/v1
-kind: Service
-metadata:
-  name: "chainlit-chatbot-sidecar"
-  labels:
-    cloud.googleapis.com/location: "europe-west1"
-  annotations:
-    run.googleapis.com/launch-stage: BETA
-    run.googleapis.com/description: side-car implementation for kounafonia chatbot only with ollama alongside
-    run.googleapis.com/ingress: all
-spec:
-  template:
-    metadata:
-      annotations:
-        run.googleapis.com/container-dependencies: "{chatbot: [ollama]}"
-    spec:
-      containers:
-        - image: europe-west1-docker.pkg.dev/lonytech/kounafonia/dev-chatbot-app:latest
-          name: chatbot
-          ports:
-            - name: http1
-              containerPort: 8000
-          resources:
-            limits:
-              cpu: 1000m
-              memory: 256Mi
-          startupProbe:
-            timeoutSeconds: 240
-            periodSeconds: 240
-            failureThreshold: 1
-            tcpSocket:
-              port: 8000
-
-        - image: ollama/ollama
-          name: ollama
-          env:
-            - name: PORT
-              value: "11434"
-
-         #  entrypoint: ["ollama", "serve"]
-          resources:
-            limits:
-              cpu: 1000m
-              memory: 512Mi
-#          volumeMounts:
-#            - name: kounafonia-storage
-#              readOnly: true
-#              mountPath: /root/
-          startupProbe:
-            timeoutSeconds: 240
-            periodSeconds: 240
-            failureThreshold: 1
-            tcpSocket:
-              port: 11434
-#      volumes:
-#        - name: kounafonia-storage # ollama-models-storage
-#          csi:
-#            driver: gcsfuse.run.googleapis.com
-#            readOnly: true
-#            volumeAttributes:
-#              bucketName: gs://kounafonia-storage
-        # [END cloudrun_mc_hello_sidecar_step_sidecar]
-      # Named volume pointing to
-      # nginx_config secret in secret manager
-      # [START cloudrun_mc_hello_sidecar_step_secret]
-#      volumes:
-#        - name: nginx-conf-secret
-#          secret:
-#            secretName: nginx_config_chat
-#            items:
-#              - key: latest
-#                path: default.conf
-      # [END cloudrun_mc_hello_sidecar_step_secret]
